Add Node Traversal Caching
Before the tree is traversed, a chunk must first initialize a Traversal, and then use that Traversal for sampling. Then the Traversal is to be disposed. The Traversal allocates the needed memory for node traversal caches. Additionally, the tree is simplified when it is saved.

Add Node Traversal Caching

Before the tree is traversed, a chunk must first initialize a Traversal, and then use that Traversal for sampling. Then the Traversal is to be disposed. The Traversal allocates the needed memory for node traversal caches. Additionally, the tree is simplified when it is saved.

Add Node Traversal Catching (SLOW)

Before the tree is traversed, a chunk must first initialize a Traversal, and then use that Traversal for sampling. Then the Traversal is to be disposed. The Traversal allocates the needed memory for node traversal caches. This is currently very slow.